Separate screen almost functional
