+++ package.json
@@ -22,7 +22,6 @@
     "react-native": "https://github.com/expo/react-native/archive/sdk-32.0.0.tar.gz",
     "react-native-calendars": "^1.22.0",
     "react-native-collapsible": "^1.4.0",
-    "react-native-datepicker": "^1.7.2",
     "react-native-elements": "^1.0.0",
     "react-native-maps": "^0.23.0",
     "react-native-modal-datetime-picker": "^6.0.0",
